requirements.txt
Flask==1.1.2
Celery==4.4.7
RabbitMQ==3.8.9
PostgreSQL==12.4
psycopg2-binary==2.8.6
requests==2.24.0
beautifulsoup4==4.9.3
lxml==4.6.2
Pillow==8.0.1
pytest==6.1.2
pytest-cov==2.10.1
Flask-JWT-Extended==3.24.1
bcrypt==3.2.0
React==17.0.1
axios==0.21.1
react-router-dom==5.2.0
redux==4.0.5
redux-thunk==2.3.0
react-redux==7.2.2
react-bootstrap==1.4.0
bootstrap==4.5.3
react-icons==3.11.0
react-toastify==6.0.8
react-chartjs-2==2.11.1
chart.js==2.9.4
react-data-table-component==6.11.6
react-csv==2.0.3
react-dropzone==11.2.4
react-image-lightbox==5.1.1
react-paginate==6.5.0
react-spinners==0.9.0
react-confirm-alert==2.6.2
react-dom==17.0.1
react-scripts==4.0.0
web-vitals==0.2.4
@testing-library/jest-dom==5.11.6
@testing-library/react==11.2.2
@testing-library/user-event==12.1.10
